+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `HashMap`
 - `Bytes.chars`
-
+- `Iterator.collect_into`
 
 ## 1.1.0 - 27/06/2024